requirements.txt
Django==1.6
Authomatic==0.0.8
Pygments==1.6
South==1.0
Unipath==1.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-admin-bootstrapped==1.6.5
django-allauth==0.17.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-extensions==1.3.8
facebook-sdk==0.4.0
gunicorn==19.0.0
oauthlib==0.6.3
pystache==0.5.4
python-decouple==2.2
python-openid==2.2.5
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
sqlformatter==1.0
sqlparse==0.1.11
static==1.0.2
wsgiref==0.1.2
psycopg2==2.5.1
django-endless-pagination==2.0
